% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abtest.R
\name{abtest_shiny}
\alias{abtest_shiny}
\title{A/B testing interactive}
\usage{
abtest_shiny(
  size_a = 100,
  size_b = 100,
  success_a = 10,
  success_b = 20,
  success_unit = "percent",
  sign_level = 0.05
)
}
\arguments{
\item{size_a}{Size of Group A}

\item{size_b}{Size of Group B}

\item{success_a}{Success of Group A}

\item{success_b}{Success of Group B}

\item{success_unit}{"count" | "percent"}

\item{sign_level}{Significance Level (typical 0.01/0.05/0.10)}
}
\description{
Launches a shiny app to A/B test
}
\examples{

# Only run examples in interactive R sessions
if (interactive())  {
   abtest_shiny()
}
}
