% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-var.R
\name{add_var_id}
\alias{add_var_id}
\title{Add a variable id at first column in dataset}
\usage{
add_var_id(data, name = "id", overwrite = FALSE)
}
\arguments{
\item{data}{A dataset}

\item{name}{Name of new variable (as string)}

\item{overwrite}{Can new id variable overwrite an existing variable in dataset?}
}
\value{
Dataset containing new id variable
}
\description{
Add a variable id at first column in dataset
}
\examples{
add_var_id(iris)
}
