% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-var.R
\name{add_var_random_01}
\alias{add_var_random_01}
\title{Add a random 0/1 variable to dataset}
\usage{
add_var_random_01(
  data,
  name = "random_01",
  prob = c(0.5, 0.5),
  overwrite = TRUE,
  seed
)
}
\arguments{
\item{data}{A dataset}

\item{name}{Name of new variable (as string)}

\item{prob}{Vector of probabilities}

\item{overwrite}{Can new random variable overwrite an existing variable in dataset?}

\item{seed}{Seed for random number generation (integer)}
}
\value{
Dataset containing new random variable
}
\description{
Add a random 0/1 variable to dataset
}
\examples{
add_var_random_01(iris)
add_var_random_01(iris, name = "my_var")
}
