% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-var.R
\name{add_var_random_starsign}
\alias{add_var_random_starsign}
\title{Add a random starsign variable to dataset}
\usage{
add_var_random_starsign(
  data,
  name = "random_starsign",
  lang = "en",
  overwrite = TRUE,
  seed
)
}
\arguments{
\item{data}{A dataset}

\item{name}{Name of new variable (as string)}

\item{lang}{Language used for starsign (en = English, de = Deutsch, es = Espanol)}

\item{overwrite}{Can new random variable overwrite an existing variable in dataset?}

\item{seed}{Seed for random number generation (integer)}
}
\value{
Dataset containing new random variable
}
\description{
Add a random starsign variable to dataset
}
\examples{
add_var_random_starsign(iris)
}
