% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunctions.R
\name{aurocEXPLORE}
\alias{aurocEXPLORE}
\title{aucrocExplore}
\usage{
aurocEXPLORE(output_path, train_data, settings_path, file_name, ...)
}
\arguments{
\item{output_path}{A string declaring the path to the settings}

\item{train_data}{Train data}

\item{settings_path}{A string declaring the path to the settings}

\item{file_name}{A string declaring the the path to the file name}

\item{...}{List of arguments}
}
\value{
auroc
}
\description{
aucrocExplore
}
