% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlyversion.R
\name{bar}
\alias{bar}
\title{Creates an interactive barchart.}
\usage{
bar(data, x)
}
\arguments{
\item{data}{A data frame}

\item{x}{A categorical variable from the data frame}
}
\value{
An interactive barchart
}
\description{
Creates an interactive barchart.
}
