% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlyversion.R
\name{bar2}
\alias{bar2}
\title{Creates an interactive barchart.}
\usage{
bar2(data, x, ...)
}
\arguments{
\item{data}{A data frame}

\item{x}{A categorical variable from the data frame}

\item{...}{Additional arguments passed onto ggplot2::geom_bar()}
}
\value{
An interactive barchart
}
\description{
Creates an interactive barchart.
}
