% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlyversion.R
\name{box}
\alias{box}
\title{Creates an interactive boxplot.}
\usage{
box(data, x)
}
\arguments{
\item{data}{A data frame}

\item{x}{A numeric variable from the data frame}
}
\value{
An interactive boxplot
}
\description{
Creates an interactive boxplot.
}
