% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunctions.R
\name{candidateNumberExplore}
\alias{candidateNumberExplore}
\title{Return the number of candidate rules for EXPLORE}
\usage{
candidateNumberExplore(OutputFile)
}
\arguments{
\item{OutputFile}{output file = paste0(output_path, file_name, ".result")}
}
\description{
Return the number of candidate rules for EXPLORE
}
