% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorexp.R
\name{cat_1}
\alias{cat_1}
\title{Generates visualizations for the distribution of one categorical variable.
Given a dataframe and the name of one categorical variable in that dataframe, this
function generates interactive plots showing the distribution of that variable.}
\usage{
cat_1(data, x)
}
\arguments{
\item{data}{A data frame}

\item{x}{A categorical variable from the data frame}
}
\value{
A plot containing several subplots
\itemize{
\item A bar chart
\item A pie chart
Note: both plots contain hover information
}
}
\description{
Generates visualizations for the distribution of one categorical variable.
Given a dataframe and the name of one categorical variable in that dataframe, this
function generates interactive plots showing the distribution of that variable.
}
