% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{changeSetting}
\alias{changeSetting}
\title{Change setting parameter value}
\usage{
changeSetting(settings, parameter, input, default_setting)
}
\arguments{
\item{settings}{Settings parameter, string}

\item{parameter}{Parameter value, string}

\item{input}{Input value}

\item{default_setting}{Default setting}
}
\value{
A setting parameter value
}
\description{
Change setting parameter value
}
