% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop.R
\name{check_vec_low_variance}
\alias{check_vec_low_variance}
\title{Check vector for low variance}
\usage{
check_vec_low_variance(values, max_prop = 0.99)
}
\arguments{
\item{values}{Vector of values}

\item{max_prop}{Maximum proportion of values without variance}
}
\value{
TRUE/FALSE (low variance)
}
\description{
Check vector for low variance
}
\examples{
\dontrun{
values <- c(1, rep(0 ,1000))
check_vec_low_variance(values, max_prop = 0.9)
}
}
