% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{clean_var}
\alias{clean_var}
\title{Clean variable}
\usage{
clean_var(
  data,
  var,
  na = NA,
  min_val = NA,
  max_val = NA,
  max_cat = NA,
  rescale01 = FALSE,
  simplify_text = FALSE,
  name = NA
)
}
\arguments{
\item{data}{A dataset}

\item{var}{Name of variable}

\item{na}{Value that replaces NA}

\item{min_val}{All values < min_val are converted to min_val (var numeric or character)}

\item{max_val}{All values > max_val are converted to max_val (var numeric or character)}

\item{max_cat}{Maximum number of different factor levels for categorical variable (if more, .OTHER is added)}

\item{rescale01}{Rescale into value between 0 and 1 (var must be numeric)}

\item{simplify_text}{if TRUE, a character variable is simplified (trim, upper, ...)}

\item{name}{New name of variable (as string)}
}
\value{
Dataset
}
\description{
Clean variable (replace NA values, set min_val and max_val)
}
\examples{
clean_var(iris, Sepal.Width, max_val = 3.5, name = "sepal_width")
}
