% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlyversion.R
\name{color_box}
\alias{color_box}
\title{Creates interactive boxplots placed side-by-side with different colors representing different levels
of the categorical variable.}
\usage{
color_box(data, x, categorize_by, legend = T)
}
\arguments{
\item{data}{A data frame}

\item{x}{A numeric variable from the data frame}

\item{categorize_by}{A categorical variable from the data frame}

<<<<<<< HEAD
\item{lengend}{Logical. 'T' for displaying the legend and 'F' for hiding the legend. Default is 'T'}
=======
\item{legend}{Logical. 'T' for displaying the legend and 'F' for hiding the legend. Default is 'T'}
>>>>>>> 8c07eca1095379ef084a3ba8c9a1ab2d87420938
}
\value{
An interactive side-by-side colored boxplot
}
\description{
Creates interactive boxplots placed side-by-side with different colors representing different levels
of the categorical variable.
}
