% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlyversion.R
\name{color_line}
\alias{color_line}
\title{Creates an interactive linegraph with different color lines for each level of the categorical variable.}
\usage{
color_line(data, x, y, categorize_by)
}
\arguments{
\item{data}{A data frame}

\item{x}{A numeric variable from the data frame that is plotted on the x-axis}

\item{y}{A numeric variable from the data frame that is plotted on the y-axis}

\item{categorize_by}{A categorical variable from the data frame}
}
\value{
An colored interactive linegraph
}
\description{
Creates an interactive linegraph with different color lines for each level of the categorical variable.
}
