% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertData.R
\name{convertData}
\alias{convertData}
\title{Convert Data}
\usage{
convertData(dframe)
}
\arguments{
\item{dframe}{data.frame to be converted, typically psychological, delivered by superordinate functions.}
}
\value{
Returns a two column-dataframe with stacked item-values in column 1 and coding for original item-affiliation in column 2.
}
\description{
Converts the structure of a data.frame to make it suitable for ggplot-usage in stackedBars and parallelBars functions.
}
\author{
Joschka Cremers
}
