% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{count_pct}
\alias{count_pct}
\title{Adds percentage to dplyr::count()}
\usage{
count_pct(data, ...)
}
\arguments{
\item{data}{A dataset}

\item{...}{Other parameters passed to count()}
}
\value{
Dataset
}
\description{
Adds variables total and
pct (percentage) to dplyr::count()
}
\examples{
count_pct(iris, Species)
}
