% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-data.R
\name{create_data_churn}
\alias{create_data_churn}
\title{Create data churn}
\usage{
create_data_churn(
  obs = 1000,
  target_name = "churn",
  factorise_target = FALSE,
  target1_prob = 0.4,
  add_id = FALSE,
  seed = 123
)
}
\arguments{
\item{obs}{Number of observations}

\item{target_name}{Variable name of target}

\item{factorise_target}{Should target variable be factorised?}

\item{target1_prob}{Probability that target = 1}

\item{add_id}{Add an id-variable to data?}

\item{seed}{Seed for randomization (integer)}
}
\value{
A dataframe
}
\description{
Artificial data that can be used for unit-testing or teaching
}
