% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-data.R
\name{create_data_esoteric}
\alias{create_data_esoteric}
\title{Create data esoteric}
\usage{
create_data_esoteric(obs = 1000, add_id = FALSE, seed = 123)
}
\arguments{
\item{obs}{Number of observations}

\item{add_id}{Add an id-variable to data?}

\item{seed}{Seed for randomization}
}
\value{
A dataset as tibble
}
\description{
Random data that can be used for unit-testing or teaching
}
\details{
Variables in dataset:
\itemize{
\item id = Identifier
\item starsign = random starsign
\item chinese = random chinese zodiac
\item moon = random moon phase
\item blood = random blood type
\item fingers_crossed = random fingers crossed (1 = yes, 0 = no)
\item success = random success (1 = yes, 0 = no)
}
}
\examples{
create_data_esoteric(obs = 100)
}
