% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-data.R
\name{create_data_person}
\alias{create_data_person}
\title{Create data person}
\usage{
create_data_person(obs = 1000, seed = 123)
}
\arguments{
\item{obs}{Number of observations}

\item{seed}{Seed for randomization (integer)}
}
\value{
A dataframe
}
\description{
Artificial data that can be used for unit-testing or teaching
}
