% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{create_notebook_explore}
\alias{create_notebook_explore}
\title{Generate a notebook}
\usage{
create_notebook_explore(output_file = "notebook-explore.Rmd", output_dir)
}
\arguments{
\item{output_file}{Filename of the html report}

\item{output_dir}{Directory where to save the html report}
}
\description{
Generate an RMarkdown Notebook template for a report. You must provide a
output-directory (parameter output_dir). The default file-name is
"notebook-explore.Rmd" (may overwrite existing file with same name)
}
\examples{
create_notebook_explore(output_file = "explore.Rmd", output_dir = tempdir())
}
