% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{create_random_data}
\alias{create_random_data}
\title{Create random data}
\usage{
create_random_data(
  obs = 1000,
  vars = 10,
  target_name = "target_ind",
  factorise_target = FALSE,
  target1_prob = 0.5,
  add_id = TRUE,
  seed = 123
)
}
\arguments{
\item{obs}{Number of observations}

\item{vars}{Number of variables}

\item{target_name}{Variable name of target}

\item{factorise_target}{Should target variable be factorised?
(from 0/1 to facotr no/yes)?}

\item{target1_prob}{Probability that buy = 1}

\item{add_id}{Add an id-variable to data?}

\item{seed}{Seed for randomization}
}
\value{
A dataframe
}
\description{
Random data that can be used for unit-testing or demonstration
}
\details{
Variables in dataset:
\itemize{
\item id = Identifier
\item var_X = variable containing values between 0 and 100
}

Target in dataset:
\itemize{
\item target_ind (may be renamed) = random values (1 = yes, 0 = no)
}
}
