% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{cut_var}
\alias{cut_var}
\title{Cut a variable}
\usage{
cut_var(data, var, bins = 8, auto_scale = FALSE, name = NA)
}
\arguments{
\item{data}{Data frame}

\item{var}{Variable}

\item{bins}{Number of bins}

\item{auto_scale}{Use auto scaling (2\%-98\%)}

\item{name}{Name of variable containing bins}
}
\value{
Data frame
}
\description{
Cut a variable
}
