% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{cut_vec_num_avg}
\alias{cut_vec_num_avg}
\title{Cut a variable}
\usage{
cut_vec_num_avg(values, bins = 8)
}
\arguments{
\item{values}{Variable}

\item{bins}{Number of bins}
}
\value{
Data frame
}
\description{
Cut a variable
}
