% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_v024.R
\name{decrypt}
\alias{decrypt}
\title{decrypt text}
\usage{
decrypt(text, codeletters = c(toupper(letters), letters, 0:9), shift = 18)
}
\arguments{
\item{text}{a text (character)}

\item{codeletters}{a string of letters that are used for decryption}

\item{shift}{number of elements shifted}
}
\value{
decrypted text
}
\description{
decrypt text
}
\examples{
decrypt("zw336 E693v")
}
