% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_v024.R
\name{describe}
\alias{describe}
\title{describe a dataset, variable or table}
\usage{
describe(data, var, target, out = "text", ...)
}
\arguments{
\item{data}{a dataset}

\item{var}{a variable of the dataset}

\item{target}{target variable (0/1 or FALSE/TRUE)}

\item{out}{output format ("text"|"list")}
}
\value{
description as table, text or list
}
\description{
describe a dataset, variable or table (depending on input parameters)
}
\examples{
describe(iris)
describe(iris, Species)
describe(iris, Sepal.Length)

iris$is_virginica <- ifelse(iris$Species == "virginica", 1, 0)
describe(iris, target = is_virginica)
}
