% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_v024.R
\name{describe_all}
\alias{describe_all}
\title{describe all variables of a dataset}
\usage{
describe_all(data = NA, out = "large")
}
\arguments{
\item{data}{a dataset}

\item{out}{output format ("small"|"large")}
}
\value{
dataset
}
\description{
describe all variables of a dataset
}
\examples{
describe_all(iris)
}
