% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_v024.R
\name{describe_cat}
\alias{describe_cat}
\title{describe categorial variable}
\usage{
describe_cat(data, var, max_cat = 10, out = "text", margin = 0)
}
\arguments{
\item{data}{a dataset}

\item{var}{variable or variable name}

\item{max_cat}{maximum number of categories displayed}

\item{out}{output format ("text"|"list")}

\item{margin}{left margin for text output (number of spaces)}
}
\value{
description as text or list
}
\description{
describe categorial variable
}
\examples{
describe_cat(iris, Species)
}
