% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_v024.R
\name{describe_num}
\alias{describe_num}
\title{describe numerical variable}
\usage{
describe_num(data, var, out = "text", margin = 0)
}
\arguments{
\item{data}{a dataset}

\item{var}{variable or variable name}

\item{out}{output format ("text"|"list")}

\item{margin}{left margin for text output (number of spaces)}
}
\value{
description as text or list
}
\description{
describe numerical variable
}
\examples{
describe_num(iris, Sepal.Length)
}
