% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_v024.R
\name{describe_tbl}
\alias{describe_tbl}
\title{describe table}
\usage{
describe_tbl(data, target, out = "text")
}
\arguments{
\item{data}{a dataset}

\item{target}{target variable}

\item{out}{output format ("text"|"list")}
}
\value{
description as text or list
}
\description{
describe table (e.g. number of rows and columns of dataset)
}
\examples{
describe_tbl(iris)

iris$is_virginica <- ifelse(iris$Species == "virginica", 1, 0)
describe_tbl(iris, is_virginica)
}
