% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divideByPattern.R
\name{divideByPattern}
\alias{divideByPattern}
\title{Divide by Pattern}
\usage{
divideByPattern(n.col = 15)
}
\arguments{
\item{dframe}{data.frame, typically psychological.}

\item{pattern}{selects those items which match the specified pattern. May also be column-index-values.}

\item{n.col}{specify a maximum number of columns, that shall be returned.}
}
\value{
Returns a data.frame which contains only the selected columns up to the maximum of n.col columns.
}
\description{
Selects only those columns of a dataframe, which match the given pattern.
Optionally the pattern is an index-value after which columns are selected.
}
\author{
Joschka Cremers
}
