% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop.R
\name{drop_obs_if}
\alias{drop_obs_if}
\title{Drop all observations where expression is true}
\usage{
drop_obs_if(data, expr)
}
\arguments{
\item{data}{Data frame}

\item{expr}{Expression}
}
\value{
Data frame
}
\description{
Drop all observations where expression is true
}
\examples{
drop_obs_if(iris, Species == "setosa")
drop_obs_if(iris, Sepal.Length < 5 | Sepal.Length >7)
}
