% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain-xgboost.R
\name{drop_obs_with_na}
\alias{drop_obs_with_na}
\title{Drop all observations with NA-values}
\usage{
drop_obs_with_na(data)
}
\arguments{
\item{data}{Data frame}
}
\value{
Data frame
}
\description{
Drop all observations with NA-values
}
