% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop.R
\name{drop_var_by_names}
\alias{drop_var_by_names}
\title{Drop variables by name}
\usage{
drop_var_by_names(data, var_names)
}
\arguments{
\item{data}{Data frame}

\item{var_names}{Vector of variable names (as string)}
}
\value{
Data frame
}
\description{
Drop variables by name
}
\examples{
drop_var_by_names(iris, "Species")
drop_var_by_names(iris, c("Sepal.Length", "Sepal.Width"))
}
