% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop.R
\name{drop_var_low_variance}
\alias{drop_var_low_variance}
\title{Drop all variables with low variance}
\usage{
drop_var_low_variance(data, max_prop = 0.99)
}
\arguments{
\item{data}{Data frame}

\item{max_prop}{Maximum proportion of values without variance}
}
\value{
Data frame
}
\description{
Drop all variables with low variance
}
\examples{
data <- data.frame(a = 1:100, b = c(0, rep(1, 99)))
drop_var_low_variance(data, max_prop = 0.9)
}
