% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop.R
\name{drop_var_no_variance}
\alias{drop_var_no_variance}
\title{Drop all variables with no variance}
\usage{
drop_var_no_variance(data)
}
\arguments{
\item{data}{Data frame}
}
\value{
Data frame
}
\description{
Drop all variables with no variance
}
\examples{
data <- data.frame(a = 1:10, b = rep(1,10))
drop_var_no_variance(data)
}
