% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain-xgboost.R
\name{drop_var_non_numeric}
\alias{drop_var_non_numeric}
\title{Drop all non numeric variables}
\usage{
drop_var_non_numeric(data)
}
\arguments{
\item{data}{Data frame}
}
\value{
Data frame
}
\description{
Drop all non numeric variables
}
