% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop.R
\name{drop_var_not_numeric}
\alias{drop_var_not_numeric}
\title{Drop all not numeric variables}
\usage{
drop_var_not_numeric(data)
}
\arguments{
\item{data}{Data frame}
}
\value{
Data frame
}
\description{
Drop all not numeric variables
}
\examples{
data <- data.frame(a = 1:10, b = rep("A",10))
drop_var_not_numeric(data)
}
