% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop.R
\name{drop_var_with_na}
\alias{drop_var_with_na}
\title{Drop all variables with NA-values}
\usage{
drop_var_with_na(data)
}
\arguments{
\item{data}{Data frame}
}
\value{
Data frame
}
\description{
Drop all variables with NA-values
}
\examples{
data <- data.frame(a = 1:10, b = rep(NA,10))
drop_var_with_na(data)
}
