% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_v024.R
\name{dwh_connect}
\alias{dwh_connect}
\title{connect to DWH}
\usage{
dwh_connect(dns = "DWH1", user = NA, pwd = NA, pwd_crypt = FALSE)
}
\arguments{
\item{dns}{DNS string}

\item{user}{user name}

\item{pwd}{password of user}

\item{pwd_crypt}{is password encryption used?}
}
\value{
connection
}
\description{
connect to datawarehouse (DWH) using ODBC
}
\examples{
con <- dwh_connect(dns = "DWH1", user = "u12345")
}
