% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_v024.R
\name{dwh_disconnect}
\alias{dwh_disconnect}
\title{disconnect from DWH}
\usage{
dwh_disconnect(connection)
}
\arguments{
\item{connection}{channel (ODBC connection)}
}
\description{
disconnect from datawarehouse (DWH) using a ODBC connection
}
\examples{
dwh_disconnect(con)
}
