% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{dwh_fastload}
\alias{dwh_fastload}
\title{write data to a DWH table}
\usage{
dwh_fastload(data, dsn, table, ...)
}
\arguments{
\item{data}{dataframe}

\item{dsn}{DSN string}

\item{table}{table name (character string)}

\item{...}{Further arguments to be passed to DBI::dbConnect()}
}
\value{
status
}
\description{
write data fast to a DWH table using a ODBC connection
Function uses packages DBI/odbc to write data faster than RODBC
Connects, writes data and disconnects
}
\examples{
\dontrun{
dwh_fastload(data, "DWH", "database.table_test")
}
}
