% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwh.R
\name{dwh_read_data}
\alias{dwh_read_data}
\title{read data from DWH}
\usage{
dwh_read_data(connection, sql, names_lower = TRUE, ...)
}
\arguments{
\item{connection}{DWH connection}

\item{sql}{sql (character string)}

\item{names_lower}{convert field names to lower (default = TRUE)}

\item{...}{Further arguments to be passed to DBI::dbGetQuery()}
}
\value{
dataframe containing table data
}
\description{
read data from DWH using a ODBC connection
}
\examples{
\dontrun{
dwh_read_data(con, "select * from database.table_test")
}
}
