% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{dwh_read_table}
\alias{dwh_read_table}
\title{read a table from DWH}
\usage{
dwh_read_table(connection, table, names_lower = TRUE, ...)
}
\arguments{
\item{connection}{DWH connection}

\item{table}{table name (character string)}

\item{names_lower}{convert field names to lower (default = TRUE)}

\item{...}{Further arguments to be passed to DBI::dbGetQuery()}
}
\value{
dataframe containing table data
}
\description{
read a table from DWH using a ODBC connection
}
\examples{
\dontrun{
dwh_read_table(con, "database.table_test")
}
}
