% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{dwh_write_table}
\alias{dwh_write_table}
\title{write data to a DWH table}
\usage{
dwh_write_table(data, connection, table, append = FALSE, rownames = FALSE,
  ...)
}
\arguments{
\item{data}{dataframe}

\item{connection}{DWH connection}

\item{table}{table name (character string)}

\item{append}{append to existing table? (default = FALSE)}
}
\value{
status
}
\description{
write data to a DWH table using a ODBC connection
}
\examples{
dwh_read_table(con, "database.table_test")
}
