% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_v024.R
\name{encrypt}
\alias{encrypt}
\title{encrypt text}
\usage{
encrypt(text, codeletters = c(toupper(letters), letters, 0:9), shift = 18)
}
\arguments{
\item{text}{a text (character)}

\item{codeletters}{a string of letters that are used for encryption}

\item{shift}{number of elements shifted}
}
\value{
encrypted text
}
\description{
encrypt text
}
\examples{
encrypt("hello world")
}
