% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{explain_forest}
\alias{explain_forest}
\title{Explain a target using Random Forest.}
\usage{
explain_forest(data, target, ntree = 50, out = "plot", ...)
}
\arguments{
\item{data}{A dataset}

\item{target}{Target variable (binary)}

\item{ntree}{Number of trees used for Random Forest}

\item{out}{Output of the function: "plot" | "model" | "importance" | all"}

\item{...}{Further arguments (passed to randomForest function)}
}
\value{
Plot of importance (if out = "plot")
}
\description{
Explain a target using Random Forest.
}
\examples{
data <- create_data_buy()
explain_forest(data, target = buy)
}
