% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{explore}
\alias{explore}
\title{explore a dataset or variable}
\usage{
explore(data, var, target, density, out = "single", ...)
}
\arguments{
\item{data}{a dataset}

\item{var}{a variable}

\item{target}{target variable (0/1 or FALSE/TRUE)}

\item{density}{using density for histograms}

\item{out}{plot layout ("single"|"double"|"all")}

\item{min_val}{all values < min_val are converted to min_val}

\item{max_val}{all values > max_val are converted to max_val}

\item{autoscale}{use 0.2 and 0.98 quantile for min_val and max_val (if min_val and max_val are not defined)}

\item{na}{value to replace NA}
}
\value{
plot object
}
\description{
explore a dataset or variable
}
\examples{
explore(iris, Species)
explore(iris, Sepal.Length)
explore(iris, Sepal.Length, density = FALSE)

iris$is_virginica <- ifelse(iris$Species == "virginica", 1, 0)
explore(iris, Species, target = is_virginica)
explore(iris, Sepal.Length, target = is_virginica)
}
