% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{explore.single_cell}
\alias{explore.single_cell}
\title{create peri-event histogram and raster plot for single_cell data}
\usage{
\method{explore}{single_cell}(
  single_cell,
  xaxis = "Time",
  yaxis = "Trials",
  stim = "Stimulus",
  shade_on = 0,
  shade_off = 0,
  shade_color = "pink"
)
}
\arguments{
\item{single_cell}{a \code{\link{single_cell}} object that has first column the spiking data and second column labeling the laps, trials or neurons}

\item{xaxis}{the x axis label of the graph, default value as "Time"}

\item{yaxis}{the y axis label of the raster plot, default value as "Trials"}

\item{stim}{The name of the highlighted region, default value as "Stimulus"}

\item{shade_on}{The starting x axis value for highlight/shading, default 0}

\item{shade_off}{The ending x axis value for highlight/shading, default 0}

\item{shade_color}{The coloring of the shading, default value as "pink"}
}
\value{
a plot with raster plot of neurons firing patterns and histogram of counts of the firing
}
\description{
create peri-event histogram and raster plot for single_cell data
}
