% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_v024.R
\name{explore_all}
\alias{explore_all}
\title{explore all variables}
\usage{
explore_all(data, target, ncol = 2, density = TRUE,
  legend_position = "non")
}
\arguments{
\item{data}{a dataset}

\item{target}{target variable (0/1 or FALSE/TRUE)}

\item{ncol}{layout of plots (number of columns)}

\item{density}{use density for histogramms}

\item{legend_position}{position of legend ("right"|"bottom"|"non")}
}
\value{
description as table, text or list
}
\description{
explore all variables of a dataset (create plots)
}
\examples{
explore_all(iris)

iris$is_virginica <- ifelse(iris$Species == "virginica", 1, 0)
explore_all(iris, target = is_virginica)
}
