% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{explore_bintar}
\alias{explore_bintar}
\title{Explore variable + binary target (values 0/1)}
\usage{
explore_bintar(data, var, target = NULL, title = NULL)
}
\arguments{
\item{data}{A dataset}

\item{var}{Numerical variable}

\item{target}{Target variable (0/1 or FALSE/TRUE)}

\item{title}{Title of the plot}
}
\value{
Plot object
}
\description{
Create a plot to explore relation between a variable and a binary target.
The target variable is choosen automatically if possible (name starts with 'target')
}
\examples{
iris$target01 <- ifelse(iris$Species == "versicolor",1,0)
explore_bintar(iris)
}
