% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_v024.R
\name{explore_cat}
\alias{explore_cat}
\title{explore categorial variable}
\usage{
explore_cat(data, var_cat, flip = TRUE, percent = TRUE, color = "#cccccc",
  auto_scale = TRUE, max_cat = 30)
}
\arguments{
\item{data}{a dataset}

\item{var_cat}{name of numerical variable}

\item{flip}{should plot be flipped? (change of x and y)}

\item{percent}{plot values as percentage (instead of absolute numbers)}

\item{color}{color of plot}

\item{max_cat}{maximum number of categories to be plotted}

\item{autoscale}{use 0.02 and 0.98 quantile for min_val and max_val (if min_val and max_val are not defined)}
}
\value{
plot object (bar chart)
}
\description{
create a plot to explore categorial variable
}
\examples{
explore_cat(iris, "Species")
}
