% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{explore_count}
\alias{explore_count}
\title{Explore count data (categories + frequency)}
\usage{
explore_count(
  data,
  cat,
  n,
  target,
  pct = FALSE,
  split = TRUE,
  title = NA,
  flip = NA
)
}
\arguments{
\item{data}{A dataset (categories + frequency)}

\item{cat}{Numerical variable}

\item{n}{Number of observations (frequency)}

\item{target}{Target variable}

\item{pct}{Show as percent?}

\item{split}{Split by target?}

\item{title}{Title of the plot}

\item{flip}{Flip plot? (for categorical variables)}
}
\value{
Plot object
}
\description{
Create a plot to explore count data (categories + freuency)
Variable named 'n' is auto detected as Frequency
}
\examples{
library(dplyr)
iris \%>\%
  count(Species) \%>\%
  explore_count(Species)
}
