% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_v024.R
\name{explore_density}
\alias{explore_density}
\title{explore density of variable}
\usage{
explore_density(data, var, target, min_val = NA, max_val = NA,
  color = "#7f7f7f", auto_scale = TRUE, ...)
}
\arguments{
\item{data}{a dataset}

\item{min_val}{all values < min_val are converted to min_val}

\item{max_val}{all values > max_val are converted to max_val}

\item{color}{color of plot}

\item{var_num}{name of numerical variable}

\item{var_target}{name of target variable (0/1 or FALSE/TRUE)}

\item{autoscale}{use 0.02 and 0.98 percent quantile for min_val and max_val (if min_val and max_val are not defined)}
}
\value{
plot object (density plot)
}
\description{
create a density plot to explore numerical variable
}
\examples{
explore_density(iris, "Sepal.Length")
iris$is_virginica <- ifelse(iris$Species == "virginica", 1, 0)
explore_density(iris, Sepal.Length, target = is_virginica)
}
