% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_v024.R
\name{explore_num}
\alias{explore_num}
\title{explore numerical variable}
\usage{
explore_num(data, var_num, min_val = NA, max_val = NA, flip = FALSE,
  color = "#cccccc", bins = 15, auto_scale = TRUE)
}
\arguments{
\item{data}{a dataset}

\item{var_num}{name of numerical variable}

\item{min_val}{all values < min_val are converted to min_val}

\item{max_val}{all values > max_val are converted to max_val}

\item{flip}{should plot be flipped? (change of x and y)}

\item{color}{color of plot}

\item{bins}{number of bins used for histogram}

\item{autoscale}{use 0.02 and 0.98 quantile for min_val and max_val (if min_val and max_val are not defined)}
}
\value{
plot object (histogram)
}
\description{
create a plot to explore numerical variable
}
\examples{
explore_num(iris, "Sepal.Length")
}
