% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{explore_points}
\alias{explore_points}
\title{Plot spatial data frames with class "POINT"
Given a spatial data frame will produce a list of different types of visualizations.}
\usage{
explore_points(data, variable = NULL)
}
\arguments{
\item{dataframe}{with geometry of "point" class}
}
\value{
A list of visualizations interactive visualizations accessible through exp. 'list$map'
}
\description{
Plot spatial data frames with class "POINT"
Given a spatial data frame will produce a list of different types of visualizations.
}
