% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{explore_polygons}
\alias{explore_polygons}
\title{Plot spatial data frames with class "POLYGON" or "MULTIPOLYGON"
Given a spatial data frame will produce a list of different types of visualizations.}
\usage{
explore_polygons(base_map_type = "OpenStreetMap", var_column = NULL)
}
\arguments{
\item{var_column}{optional variable to plot a scaled and/or gradient map}

\item{dataframe}{with geometry of "POLYGON" or "MULTIPOLYGON" class}
}
\value{
A list of visualizations interactive visualizations accessible through exp. 'list$map'
}
\description{
Plot spatial data frames with class "POLYGON" or "MULTIPOLYGON"
Given a spatial data frame will produce a list of different types of visualizations.
}
