% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_v024.R
\name{explore_shiny}
\alias{explore_shiny}
\title{explore dataset interactive}
\usage{
explore_shiny(data, target)
}
\arguments{
\item{data}{a dataset}

\item{target}{target variable (0/1 or FALSE/TRUE)}
}
\description{
launches a shiny app to explore a dataset
}
\examples{
explore_shiny(iris)
}
