% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{explore_tbl}
\alias{explore_tbl}
\title{Explore table}
\usage{
explore_tbl(data)
}
\arguments{
\item{data}{A dataset}
}
\description{
Explore a table. Plots variable types, variables with no variance and variables with NA
}
\examples{
explore_tbl(iris)
}
