% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlyversion.R
\name{facet_bar}
\alias{facet_bar}
\title{Creates an interactive barchart faceted by a categorical variable.}
\usage{
facet_bar(data, x, y, categorize_by)
}
\arguments{
\item{data}{A data frame}

\item{x}{A numeric variable from the data frame that is plotted on the x-axis}

\item{y}{A numeric variable from the data frame that is plotted on the y-axis}

\item{categorize_by}{A categorical variable from the data frame}
}
\value{
An interactive barchart faceted by a categorical variable
}
\description{
Creates an interactive barchart faceted by a categorical variable.
}
