% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_v024.R
\name{format_num}
\alias{format_num}
\title{format number}
\usage{
format_num(number = 0, digits = 1)
}
\arguments{
\item{number}{a number (integer or real)}

\item{digits}{number of digits}
}
\value{
formated number as text
}
\description{
formats a big number as k (1000) or M (100000)
}
\examples{
format_num(5500, digits = 2)
}
