% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{format_num_auto}
\alias{format_num_auto}
\title{Format number as character string (auto)}
\usage{
format_num_auto(number = 0, digits = 1)
}
\arguments{
\item{number}{A number (integer or real)}

\item{digits}{Number of digits}
}
\value{
Formated number as text
}
\description{
Formats a number depending on the value as number with space, scientific or
big number as k (1 000), M (1 000 000) or B (1 000 000 000)
}
\examples{
format_num_kMB(5500, digits = 2)
}
