% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_v024.R
\name{format_target}
\alias{format_target}
\title{format target}
\usage{
format_target(target)
}
\arguments{
\item{target}{attribute as vector}
}
\value{
formated target
}
\description{
formats a target as a 0/1 attribute
}
\examples{
data$target <- format_target(data$target)
}
