% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_v024.R
\name{format_type}
\alias{format_type}
\title{format type description}
\usage{
format_type(type)
}
\arguments{
\item{type}{type description ("integer", "double", "logical", character")}
}
\value{
formated type description (int|dou|log|chr)
}
\description{
format type description of varable to 3 letters (int|dou|log|chr)
}
\examples{
format_type(typeof(iris$Species))
}
