% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_v024.R
\name{get_nrow}
\alias{get_nrow}
\title{get number of rows for a grid plot}
\usage{
get_nrow(varnames, exclude = 0, ncol = 2)
}
\arguments{
\item{varnames}{list of variables to be plotted}

\item{exclude}{number of variables that will be excluded from plot}
}
\value{
number of rows
}
\description{
get number of rows for a grid plot
}
\examples{
get_nrow(names(iris), ncol = 2)
}
