% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_v024.R
\name{get_type}
\alias{get_type}
\title{return type of variable}
\usage{
get_type(var)
}
\arguments{
\item{var}{a vector (dataframe column)}
}
\value{
value of typeof or "other"
}
\description{
return value of typeof, except if variable contains <hide>, then return "other"
}
\examples{
get_type(iris$Species)
}
