% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{get_var_buckets}
\alias{get_var_buckets}
\title{Put variables into "buckets" to create a set of plots instead one large plot}
\usage{
get_var_buckets(data, bucket_size = 100, target_var_name = NA)
}
\arguments{
\item{data}{A dataset}

\item{bucket_size}{Maximum number of variables in one bucket}

\item{target_var_name}{Name of the target variable (if defined)}
}
\value{
Buckets as a list
}
\description{
Put variables into "buckets" to create a set of plots instead one large plot
}
\examples{
total_fig_height(iris)
total_fig_height(iris, target = Species)
total_fig_height(nvar = 5)
}
