% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{getSetting}
\alias{getSetting}
\title{Get the setting parameter value}
\usage{
getSetting(settings, parameter, type = "value")
}
\arguments{
\item{settings}{}

\item{parameter}{}

\item{type}{}
}
\value{
A parameter value, character.
}
\description{
Get the setting parameter value
}
