% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlyversion.R
\name{group_box}
\alias{group_box}
\title{Creates interactive boxplots placed side-by-side with different colors representing different levels
of one categorical variable, split by the levels of another categorical variable on the x-axis.}
\usage{
group_box(data, x_cat, y, group)
}
\arguments{
\item{data}{A data frame}

\item{x_cat}{A categorical variable from the data frame that is plotted on the x-axis}

\item{y}{A numeric variable from the data frame that is plotted on the y-axis}

\item{group}{A categorical variable from the data frame that is represented through color}
}
\value{
An interactive grouped boxplot
}
\description{
Creates interactive boxplots placed side-by-side with different colors representing different levels
of one categorical variable, split by the levels of another categorical variable on the x-axis.
}
