% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_v024.R
\name{guess_cat_num}
\alias{guess_cat_num}
\title{return if variable is categorial or nomerical}
\usage{
guess_cat_num(var)
}
\arguments{
\item{var}{a vector (dataframe column)}
}
\value{
"cat" (categorial), "num" (numerical) or "oth" (other)
}
\description{
guess if variable is categorial or numerical based on name, type and values of variable
}
\examples{
guess_cat_num(iris$Species)
}
