% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlyversion.R
\name{hist}
\alias{hist}
\title{Creates an interactive histogram.}
\usage{
hist(data, x)
}
\arguments{
\item{data}{A data frame}

\item{x}{A categorical variable from the data frame}
}
\value{
An interactive histogram
}
\description{
Creates an interactive histogram.
}
