% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{interact}
\alias{interact}
\title{Make a explore-plot interactive}
\usage{
interact(obj, lower_title = TRUE, hide_geom_text = TRUE)
}
\arguments{
\item{obj}{A object (e.g. ggplot2-object)}

\item{lower_title}{Lowering the title in ggplot2-object(\code{FALSE}/\code{TRUE})}

\item{hide_geom_text}{Hiding geom_text in ggplot2-object (\code{FALSE}/\code{TRUE})}
}
\value{
Plot object
}
\description{
Make a explore-plot interactive
}
\examples{
library(dplyr)
if (interactive())  {
   iris \%>\% explore(Sepal.Length, target = Species) \%>\% interact()
}
}
