% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlyversion.R
\name{line}
\alias{line}
\title{Creates an interactive linegraph.}
\usage{
line(data, x, y)
}
\arguments{
\item{data}{A data frame}

\item{x}{A numeric variable from the data frame that is plotted on the x-axis}

\item{y}{A numeric variable from the data frame that is plotted on the y-axis}
}
\value{
An interactive linegraph
}
\description{
Creates an interactive linegraph.
}
