% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain-xgboost.R
\name{log_info_if}
\alias{log_info_if}
\title{Log conditional}
\usage{
log_info_if(log = TRUE, text = "log")
}
\arguments{
\item{log}{log (TRUE|FALSE)}

\item{text}{text string to be logged}
}
\value{
prints log on screen (if log == TRUE).
}
\description{
Log conditional
}
