% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_function_ideas.R
\name{makeSpatialViz}
\alias{makeSpatialViz}
\title{Plot spatial data frames bases on data type exp. "MULTIPOLYGON/POLYGON", and "POINT"
Given a spatial data frame of either polygon, multipolygon, or point geometry class, will produce a list of different types of visualizations.}
\usage{
makeSpatialViz(data, variable = NULL)
}
\arguments{
\item{dataframe}{with geometry of either polygon/multipolygon or point class}

\item{optional--variable}{variable to plot a scaled and/or gradient map}
}
\value{
A list of visualizations interactive visualizations accessible through exp. 'list$map'
}
\description{
Plot spatial data frames bases on data type exp. "MULTIPOLYGON/POLYGON", and "POINT"
Given a spatial data frame of either polygon, multipolygon, or point geometry class, will produce a list of different types of visualizations.
}
\examples{
point_data <- st_sample(st_read(system.file("shape/nc.shp", package="sf")), 100)
point_data <- st_as_sf(point_data)
point_data <- st_as_sf(point_data)

}
