% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{mix_color}
\alias{mix_color}
\title{mix_color}
\usage{
mix_color(color1, color2 = NA, n = 5)
}
\arguments{
\item{n}{Number of different colors that should be generated}

\item{color}{Color that is used to generate the palette}
}
\value{
Vector of color-codes
}
\description{
mix_color
}
\examples{
mix_color("blue", n = 10)
mix_color("gold", "red", n = 4)
}
