% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunctions.R
\name{modelsCurveExplore}
\alias{modelsCurveExplore}
\title{modelsCurveExplore # TODO: update documentation?}
\usage{
modelsCurveExplore(
  train_data = NULL,
  settings_path = NULL,
  output_path,
  file_name = "train_data",
  OutputFile = NULL,
  StartRulelength = 1,
  EndRulelength = 3,
  OperatorMethod = "EXHAUSTIVE",
  CutoffMethod = "ALL",
  ClassFeature = "'class'",
  PositiveClass = "'Iris-versicolor'",
  FeatureInclude = "",
  Maximize = "BALANCEDACCURACY",
  Accuracy = 0,
  BalancedAccuracy = 0,
  Specificity = 0,
  OutputMethod = "BEST",
  PrintSettings = TRUE,
  PrintPerformance = FALSE,
  Subsumption = FALSE,
  BranchBound = TRUE,
  Sorted = "none",
  Parallel = TRUE,
  ParallelMethod = "TWO",
  BinaryReduction = FALSE
)
}
\arguments{
\item{train_data}{Train data}

\item{settings_path}{A string declaring the path to the settings}

\item{output_path}{A string declaring the path to the settings}

\item{file_name}{A string declaring the the path to the file name}

\item{...}{List of arguments}
}
\value{
models for different sensitivities/specificities
}
\description{
modelsCurveExplore # TODO: update documentation?
}
