% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{new_single_cell}
\alias{new_single_cell}
\title{Given a matlab data set, this function creates a dataframe}
\usage{
new_single_cell(x)
}
\arguments{
\item{x}{the matlab data with the first column neurons firing time/location, eg. the file path to a matlab file}
}
\description{
Given a matlab data set, this function creates a dataframe
}
