% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorexp.R
\name{num1_cat1}
\alias{num1_cat1}
\title{Generates visualizations for the distribution of one numeric variable split by
one categorical variable and vice versa.
Given a dataframe and the names of one numeric variable and one categorical variable in that dataframe, this
function generates interactive plots showing the distribution of one variable against the other.}
\usage{
num1_cat1(data, x, categorize_by, ...)
}
\arguments{
\item{data}{A data frame}

\item{x}{A numeric variable from the data frame}

\item{categorize_by}{A categorical variable from the data frame}

\item{...}{Additional arguments for ggplot2::geom_bar() and ggplot2::geom_errorbar}
}
\value{
A plot containing several subplots
\itemize{
\item A barchart with error bars
\item A stacked barchart
\item A side-by-side barchart
\item A colored boxplot
\item An overlapping histogram
Note: all plots contain hover information
}
}
\description{
Generates visualizations for the distribution of one numeric variable split by
one categorical variable and vice versa.
Given a dataframe and the names of one numeric variable and one categorical variable in that dataframe, this
function generates interactive plots showing the distribution of one variable against the other.
}
