% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorexp.R
\name{num1_cat2}
\alias{num1_cat2}
\title{Generates a visualization for the distribution of two categorical variables based on
one numeric variable.
Given a dataframe and the names of two categorical variables and one numeric variable, this
function generates an interactive plot showing the distribution of these variables against each other.}
\usage{
num1_cat2(data, x_cat, y, group, ...)
}
\arguments{
\item{data}{A data frame}

\item{x_cat}{A categorical variable from the data frame that is plotted on the x-axis}

\item{y}{A numeric variable from the data frame that is plotted on the y-axis}

\item{group}{A categorical variable from the data frame that is represented through color}
}
\value{
An interactive colored boxplot
Note: this plot contains hover information
}
\description{
Generates a visualization for the distribution of two categorical variables based on
one numeric variable.
Given a dataframe and the names of two categorical variables and one numeric variable, this
function generates an interactive plot showing the distribution of these variables against each other.
}
