% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorexp.R
\name{num2_cat1}
\alias{num2_cat1}
\title{Generates visualizations for the distribution of two numeric variables split by
one categorical variable.
Given a dataframe and the names of two numeric variables and one categorical variable, this
function generates interactive plots showing the distribution of these variables against each other.}
\usage{
num2_cat1(data, x, y, categorize_by, ...)
}
\arguments{
\item{data}{A data frame}

\item{x}{A numeric variable from the data frame that is plotted on the x-axis}

\item{y}{A numeric variable from the data frame that is plotted on the y-axis}

\item{categorize_by}{A categorical variable from the data frame}
}
\value{
A plot containing several subplots
\itemize{
\item A scatterplot with colored points corresponding to levels of the categorical variable
\item A colored linegraph
\item A faceted histogram
Note: all plots contain hover information
}
}
\description{
Generates visualizations for the distribution of two numeric variables split by
one categorical variable.
Given a dataframe and the names of two numeric variables and one categorical variable, this
function generates interactive plots showing the distribution of these variables against each other.
}
