% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorexp.R
\name{num_1}
\alias{num_1}
\title{Generates visualizations for the distribution of a numeric variable.
Given a dataframe and the name of a numeric variable in that dataframe, this
function generates interactive plots showing the distribution of that variable.}
\usage{
num_1(data, x)
}
\arguments{
\item{data}{A data frame}

\item{x}{A numeric variable from the data frame}
}
\value{
A plot containing several subplots
\itemize{
\item A boxplot with quartiles computed through the linear method
\item A histogram
Note: both plots contain hover information
}
}
\description{
Generates visualizations for the distribution of a numeric variable.
Given a dataframe and the name of a numeric variable in that dataframe, this
function generates interactive plots showing the distribution of that variable.
}
