% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorexp.R
\name{num_2}
\alias{num_2}
\title{Generates visualizations for the distribution of two numeric variables.
Given a dataframe and the names of two numeric variable in that dataframe, this
function generates interactive plots showing the distribution of one variable against the other.}
\usage{
num_2(data, x, y)
}
\arguments{
\item{data}{A data frame}

\item{x}{A numeric variable from the data frame that is plotted on the x-axis}

\item{y}{A numeric variable from the data frame that is plotted on the y-axis}
}
\value{
A plot containing several subplots
\itemize{
\item A scatterplot
\item A linegraph
Note: both plots contain hover information
}
}
\description{
Generates visualizations for the distribution of two numeric variables.
Given a dataframe and the names of two numeric variable in that dataframe, this
function generates interactive plots showing the distribution of one variable against the other.
}
