% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlyversion.R
\name{overlap_hist}
\alias{overlap_hist}
\title{Creates an overlapping histogram.}
\usage{
overlap_hist(data, x, categorize_by)
}
\arguments{
\item{data}{A data frame}

\item{x}{A numeric variable from the data frame}

\item{categorize_by}{A categorical variable from the data frame}
}
\value{
An interactive overlapping histogram
}
\description{
Creates an overlapping histogram.
}
