% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlyversion.R
\name{par_bar}
\alias{par_bar}
\title{Creates an interactive side-by-side barchart.}
\usage{
par_bar(data, x, categorize_by, legend = "right", ...)
}
\arguments{
\item{data}{A data frame}

\item{x}{A numeric variable from the data frame}

\item{categorize_by}{A categorical variable from the data frame}

<<<<<<< HEAD
=======
\item{legend}{position of the legend. Default is "right"}

>>>>>>> 8c07eca1095379ef084a3ba8c9a1ab2d87420938
\item{...}{Additional arguments passed onto ggplot2::geom_bar()}
}
\value{
An interactive side-by-side barchart
}
\description{
Creates an interactive side-by-side barchart.
}
