% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelBarCharts.R
\name{parallelBars}
\alias{parallelBars}
\title{Parallel Barchart}
\usage{
parallelBars(n.col = 15, palette = "Blues", direction = 1)
}
\arguments{
\item{dframe}{data.frame to be analyzed, typically psychological.}

\item{pattern}{optional, selects those items which match the specified pattern. May also be column-index-values.}

\item{n.col}{specify the maximum number of columns to be drawn.}

\item{palette}{specify a color-palette from RColorBrewer.}

\item{direction}{set to -1 to reverse the color-scheme.}
}
\value{
Plots parallel bars in a chart with a maximum of 15 items at the x-axis and frequencies on y-axis.
}
\description{
Draw a parallel barchart for a general overview of items from a dataset.
}
\examples{
#Load big5-data from Explore-package
data("big5")

#Draw only Extraversion-items in graph
parallelBars(big5, pattern = "^E", palette = "YlGn")

}
\author{
Joschka Cremers
}
