% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlyversion.R
\name{pie}
\alias{pie}
\title{Creates an interactive pie chart.}
\usage{
pie(data, x, ...)
}
\arguments{
\item{data}{A data frame}

\item{x}{A categorical variable from the data frame}
}
\value{
An interactive pie chart
}
\description{
Creates an interactive pie chart.
}
