% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{plot_legend_targetpct}
\alias{plot_legend_targetpct}
\title{Plots a legend that can be used for explore_all with a binary target}
\usage{
plot_legend_targetpct(border = TRUE)
}
\arguments{
\item{border}{Draw a border?}
}
\value{
Base plot
'@importFrom graphics legend par plot
}
\description{
Plots a legend that can be used for explore_all with a binary target
}
\examples{
plot_legend_targetpct(border = TRUE)
}
