% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{plot_var_info}
\alias{plot_var_info}
\title{Plot a variable info}
\usage{
plot_var_info(data, var, info = "")
}
\arguments{
\item{data}{A dataset}

\item{var}{Variable}

\item{info}{Text to plot}
}
\value{
Plot (ggplot)
}
\description{
Creates a ggplot with the variable-name as title and a text
}
\examples{
plot_var_info(iris, Species, info = "don't pick me!")
}
