% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_v024.R
\name{replace_na_with}
\alias{replace_na_with}
\title{replace NA}
\usage{
replace_na_with(data, var_name, with)
}
\arguments{
\item{data}{a dataframe}

\item{var_name}{name of variable where NAs are replaced}

\item{with}{value instead of NA}
}
\value{
updated dataframe
}
\description{
replace NA values of an attribute in a dataframe
}
\examples{
data <- data.frame(nr = c(1,2,3,NA,NA))
replace_na_with(data, "nr", 0)
}
