% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{report}
\alias{report}
\title{generate a report of all attributes}
\usage{
report(data, target, density = FALSE, output_file)
}
\arguments{
\item{data}{a dataset}

\item{target}{target variable (0/1 or FALSE/TRUE)}

\item{density}{use density?}

\item{output_file}{path/filename.html}
}
\description{
generate a report of all attributes
if targed is defined, the relation to the target is reported
}
\examples{
report(iris)
}
