% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{rescale01}
\alias{rescale01}
\title{Rescales a numeric variable into values between 0 and 1}
\usage{
rescale01(x)
}
\arguments{
\item{var}{text string}
}
\value{
text string
}
\description{
Rescales a numeric variable into values between 0 and 1
}
\examples{
simplify_text(" Hello  World !, ")
}
