% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunctions.R
\name{resultsExplore}
\alias{resultsExplore}
\title{Return a set of results from EXPLORE output file}
\usage{
resultsExplore(outputFile)
}
\arguments{
\item{outputFile}{outputfile = paste0(output_path, file_name, ".result")}
}
\description{
Return a set of results from EXPLORE output file
}
