% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunctions.R
\name{rocCurveExplore}
\alias{rocCurveExplore}
\title{rocCurveExplore}
\usage{
rocCurveExplore(modelsCurve, data, labels)
}
\arguments{
\item{modelsCurve}{modelsCurve}

\item{data}{data}

\item{labels}{labels}
}
\value{
auc value for EXPLORE
}
\description{
rocCurveExplore
}
