% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{saveData}
\alias{saveData}
\title{saveData}
\usage{
saveData(output_path, train_data, file_name)
}
\arguments{
\item{output_path}{A string declaring the path to the settings}

\item{train_data}{Train data}

\item{file_name}{A string declaring the the path to the file name}
}
\description{
saveData
}
