% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlyversion.R
\name{scatter}
\alias{scatter}
\title{Creates an interactive scatterplot.}
\usage{
scatter(data, x, y)
}
\arguments{
\item{data}{A data frame}

\item{x}{A numeric variable from the data frame that is plotted on the x-axis}

\item{y}{A numeric variable from the data frame that is plotted on the y-axis}
}
\value{
An interactive scatterplot
}
\description{
Creates an interactive scatterplot.
}
