% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunctions.R
\name{settingsExplore}
\alias{settingsExplore}
\title{settingsExplore}
\usage{
settingsExplore(
  settings,
  output_path,
  file_name,
  train_data = NULL,
  OutputFile = NULL,
  StartRulelength = NULL,
  EndRulelength = NULL,
  OperatorMethod = NULL,
  CutoffMethod = NULL,
  ClassFeature,
  PositiveClass,
  FeatureInclude = NULL,
  Maximize = NULL,
  Accuracy = NULL,
  Specificity = NULL,
  PrintSettings = NULL,
  PrintPerformance = NULL,
  Subsumption = NULL,
  BranchBound = NULL,
  Parallel = NULL
)
}
\arguments{
\item{settings}{}

\item{output_path}{}

\item{file_name}{}

\item{train_data}{}

\item{maxRuleLength}{}

\item{mandatoryFeatures}{}
}
\value{

}
\description{
settingsExplore
}
