% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{show_color}
\alias{show_color}
\title{show_color}
\usage{
show_color(color)
}
\arguments{
\item{color}{}
}
\value{
ggplot
}
\description{
show_color
}
\examples{
show_color("gold")
show_color(c("blue", "red", "green"))
}
