% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{simplify_text}
\alias{simplify_text}
\title{Simplifies a text string}
\usage{
simplify_text(text)
}
\arguments{
\item{text}{text string}
}
\value{
text string
}
\description{
A text string is converted into a simplified version by
trimming, converting to upper case, replacing german Umlaute,
dropping special characters like comma and semicolon and
replacing multiple spaces with one space.
}
\examples{
simplify_text(" Hello  World !, ")
}
