% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{single_cell}
\alias{single_cell}
\title{Create a single_cell object from a .mat file}
\usage{
single_cell(matdata)
}
\arguments{
\item{matdat}{the matlab data with the first column neurons firing time/location, eg. the file path to a matlab file}
}
\description{
Create a single_cell object from a .mat file
}
