% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlyversion.R
\name{stack_bar}
\alias{stack_bar}
\title{Creates an interactive stacked barchart.}
\usage{
stack_bar(data, x, categorize_by, legend = "right", ...)
}
\arguments{
\item{data}{A data frame}

\item{x}{A numeric variable from the data frame}

\item{categorize_by}{A categorical variable from the data frame}
}
\value{
An interactive stacked barchart
}
\description{
Creates an interactive stacked barchart.
}
