% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackedBarCharts.R
\name{stackedBars}
\alias{stackedBars}
\title{Stacked Barchart}
\usage{
stackedBars(type = position_fill, n.col = 15, palette = "Blues", direction = 1)
}
\arguments{
\item{dframe}{data.frame to be analyzed, typically psychological.}

\item{type}{Choose between relative (type = "position_fill") and absolute (type = "position_stack") frequencies.}

\item{pattern}{optional, selects those items which match the specified pattern. May also be column-index-values.}

\item{n.col}{specify the maximum number of columns to be drawn.}

\item{palette}{specify a color-palette from RColorBrewer. Advised to use bidirectional (e.g. "RdBu") and categorical (e.g. "Set3") color-schemes if apropriate.}

\item{direction}{set to -1 to reverse the color-scheme.}
}
\value{
Plots a stacked barchart with a maximum of 15 items at the y-axis and relative or absolute frequencies on x-axis.
}
\description{
Draw a stacked barchart for a general overview of items from a dataset.
}
\examples{
#Load big5-data from Explore-package
data("big5")

#Draw only Extraversion-items in graph
stackedBars(big5, pattern = "^E", palette = "YlGn")

}
\author{
Joschka Cremers
}
