% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_v024.R
\name{target_explore_cat}
\alias{target_explore_cat}
\title{explore categorial variable + target}
\usage{
target_explore_cat(data, var_cat, var_target = "target_ind", min_val = NA,
  max_val = NA, flip = TRUE, num2char = TRUE, title = NA,
  auto_scale = TRUE, max_cat = 30, legend_position = "bottom")
}
\arguments{
\item{data}{a dataset}

\item{var_cat}{name of categorial variable}

\item{var_target}{name of target variable (0/1 or FALSE/TRUE)}

\item{min_val}{all values < min_val are converted to min_val}

\item{max_val}{all values > max_val are converted to max_val}

\item{flip}{should plot be flipped? (change of x and y)}

\item{num2char}{if TRUE, numeric values in variable are converted into character}

\item{title}{title of plot}

\item{max_cat}{maximum numbers of categories to be plotted}

\item{legend_position}{position of legend ("right"|"bottom"|"non")}

\item{autoscale}{not used, just for compatibility}
}
\value{
plot object
}
\description{
create a plot to explore relation between categorial variable and a binary target
}
\examples{
iris$is_virginica <- ifelse(iris$Species == "virginica", 1, 0)
target_explore_cat(iris, "Species", "is_virginica")
}
