% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_v024.R
\name{target_explore_num}
\alias{target_explore_num}
\title{explore categorial variable + target}
\usage{
target_explore_num(data, var_num, var_target = "target_ind", min_val = NA,
  max_val = NA, flip = TRUE, title = NA, auto_scale = TRUE, na = NA,
  legend_position = "bottom")
}
\arguments{
\item{data}{a dataset}

\item{var_target}{name of target variable (0/1 or FALSE/TRUE)}

\item{min_val}{all values < min_val are converted to min_val}

\item{max_val}{all values > max_val are converted to max_val}

\item{flip}{should plot be flipped? (change of x and y)}

\item{title}{title of plot}

\item{na}{value to replace NA}

\item{legend_position}{position of legend ("right"|"bottom"|"non")}

\item{var_cat}{name of numerical variable}

\item{autoscale}{use 0.02 and 0.98 quantile for min_val and max_val (if min_val and max_val are not defined)}
}
\value{
plot object
}
\description{
create a plot to explore relation between numerical variable and a binary target
}
\examples{
iris$is_virginica <- ifelse(iris$Species == "virginica", 1, 0)
target_explore_num(iris, "Sepal.Length", "is_virginica")
}
