% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunctions.R
\name{trainExplore}
\alias{trainExplore}
\title{Finds decision rule using Explore}
\usage{
trainExplore(
  output_path,
  train_data = NULL,
  settings_path = NULL,
  file_name = "train_data",
  ...
)
}
\arguments{
\item{output_path}{A string declaring the path to the settings}

\item{train_data}{Train data}

\item{settings_path}{A string declaring the path to the settings}

\item{file_name}{A string declaring the the path to the file name}

\item{...}{A list of arguments}
}
\value{
Model
}
\description{
`trainExplore()` finds the best decision rule for the given data set based using EXPLORE.
}
