% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-data.R
\name{use_data_beer}
\alias{use_data_beer}
\title{Use the beer data set}
\usage{
use_data_beer()
}
\value{
Dataset as tibble
}
\description{
This data set is an incomplete collection of popular beers in
Austria, Germany and Switzerland. Data are collected from various
websites in 2023. Some of the collected data may be incorrect.
}
\examples{
use_data_beer()
}
