% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-data.R
\name{use_data_diamonds}
\alias{use_data_diamonds}
\title{Use the diamonds data set}
\usage{
use_data_diamonds()
}
\value{
Dataset
}
\description{
This data set comes with the ggplot2 package.
It contains the prices and other attributes of almost 54,000 diamonds.
}
\examples{
use_data_diamonds()
}
\seealso{
\code{\link[ggplot2:diamonds]{ggplot2::diamonds}}
}
