% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-data.R
\name{use_data_iris}
\alias{use_data_iris}
\title{Use the iris flower data set}
\usage{
use_data_iris()
}
\value{
Dataset as tibble
}
\description{
This data set comes with base R. The data set gives the measurements
in centimeters of the variables sepal length and width and petal length and
width, respectively, for 50 flowers from each of 3 species of iris.
The species are Iris setosa, versicolor, and virginica.
}
\examples{
use_data_iris()
}
