% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-data.R
\name{use_data_mpg}
\alias{use_data_mpg}
\title{Use the mpg data set}
\usage{
use_data_mpg()
}
\value{
Dataset
}
\description{
This data set comes with the ggplot2 package.
It contains a subset of the fuel economy data that the EPA makes available on
https://fueleconomy.gov/. It contains only models which had a new
release every year between 1999 and 2008 - this was used as a proxy for the
popularity of the car.
}
\examples{
use_data_mpg()
}
