% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-data.R
\name{use_data_mtcars}
\alias{use_data_mtcars}
\title{Use the mtcars data set}
\usage{
use_data_mtcars()
}
\value{
Dataset
}
\description{
This data set comes with base R. The data was extracted from
the 1974 Motor Trend US magazine, and comprises fuel consumption and
10 aspects of automobile design and performance for 32 automobiles
(1973–74 models).
}
\examples{
use_data_mtcars()
}
