% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-data.R
\name{use_data_penguins}
\alias{use_data_penguins}
\title{Use the penguins data set}
\usage{
use_data_penguins(short_names = FALSE)
}
\arguments{
\item{short_names}{Use short variable names}
}
\value{
Dataset
}
\description{
This data set comes with the palmerpenguins package.
It contains measurements for penguin species, island in Palmer Archipelago,
size (flipper length, body mass, bill dimensions), and sex.
}
\examples{
use_data_penguins()
use_data_penguins(short_names = TRUE)
}
\seealso{
\code{\link[palmerpenguins:penguins]{palmerpenguins::penguins}}
}
