% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-data.R
\name{use_data_starwars}
\alias{use_data_starwars}
\title{Use the starwars data set}
\usage{
use_data_starwars()
}
\value{
Dataset
}
\description{
This data set comes with the dplyr package.
It contains data of 87 star war characters
}
\examples{
use_data_starwars()
}
\seealso{
\code{\link[dplyr:starwars]{dplyr::starwars}}
}
