% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-data.R
\name{use_data_titanic}
\alias{use_data_titanic}
\title{Use the titanic data set}
\usage{
use_data_titanic(count = FALSE)
}
\arguments{
\item{count}{use count data}
}
\value{
Dataset
}
\description{
This data set comes with base R.
Survival of passengers on the Titanic.
}
\examples{
use_data_titanic()
}
