% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-data.R
\name{use_data_wordle}
\alias{use_data_wordle}
\title{Use the wordle data set}
\usage{
use_data_wordle()
}
\value{
Dataset
}
\description{
This data set contains the result of a real wordle challange (in german language)
between tow players. Wordle is a game where a player guesses a five-letter
word in six tries. The variable "try" reflects the success of player A and B.
Other variables like "noun", "aeiou", "unique", "common" and "rare" reflect
the properties of the word.
}
\examples{
use_data_wordle()
}
