% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{yyyymm_calc}
\alias{yyyymm_calc}
\title{Calculate with periods (format yyyymm)}
\usage{
yyyymm_calc(yyyymm, add_month = 0, add_year = 0)
}
\arguments{
\item{yyyymm}{Input vector of periods (format yyyymm)}

\item{add_month}{How many months to add (can be negative too)}

\item{add_year}{How many years to add (can be negative too)}
}
\value{
Vector of periods (format yyyymm)
}
\description{
Calculate with periods (format yyyymm)
}
\examples{
yyyymm_calc(202412, add_month = 1)
yyyymm_calc(c(202411,202412,202501), add_month = -1, add_year = 1)
}
