% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunctions.R
\name{aurocEXPLORE}
\alias{aurocEXPLORE}
\title{aucrocExplore}
\usage{
aurocEXPLORE(output_path, train_data, settings_path, file_name, ...)
}
\arguments{
\item{output_path}{}

\item{train_data}{}

\item{settings_path}{}

\item{file_name}{}

\item{...}{}
}
\value{
auroc
}
\description{
aucrocExplore
}
