% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{changeSetting}
\alias{changeSetting}
\title{Change setting parameter value}
\usage{
changeSetting(settings, parameter, input, default_setting)
}
\arguments{
\item{settings}{}

\item{parameter}{}

\item{input}{}

\item{default_setting}{}
}
\value{
A setting parameter value
}
\description{
Change setting parameter value
}
