% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlyversion.R
\name{color_box}
\alias{color_box}
\title{Creates interactive boxplots placed side-by-side with different colors representing different levels
of the categorical variable.}
\usage{
color_box(data, x, categorize_by, legend = T)
}
\arguments{
\item{data}{A data frame}

\item{x}{A numeric variable from the data frame}

\item{categorize_by}{A categorical variable from the data frame}
}
\value{
An interactive side-by-side colored boxplot
}
\description{
Creates interactive boxplots placed side-by-side with different colors representing different levels
of the categorical variable.
}
