% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-data.R
\name{create_data_empty}
\alias{create_data_empty}
\title{Create an empty dataset}
\usage{
create_data_empty(obs = 1000, add_id = FALSE)
}
\arguments{
\item{obs}{Number of observations}

\item{add_id}{Add an id}
}
\value{
Dataset
}
\description{
Create an empty dataset
}
\examples{
create_data_empty()
}
