% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-data.R
\name{create_data_unfair}
\alias{create_data_unfair}
\title{Create data unfair}
\usage{
create_data_unfair(obs = 1000, add_id = FALSE, seed = 123)
}
\arguments{
\item{obs}{Number of observations}

\item{add_id}{Add an id-variable to data?}

\item{seed}{Seed for randomization (integer)}
}
\value{
A dataframe
}
\description{
Artificial data that can be used for unit-testing or teaching
(fairness & AI bias)
}
