% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\title{Describe a dataset or variable}
\usage{
describe(data, var, n, target, out = "text", ...)
}
\arguments{
\item{data}{A dataset}

\item{var}{A variable of the dataset}

\item{n}{Weights variable for count-data}

\item{target}{Target variable (0/1 or FALSE/TRUE)}

\item{out}{Output format ("text"|"list") of variable description}

\item{...}{Further arguments}
}
\value{
Description as table, text or list
}
\description{
Describe a dataset or variable (depending on input parameters)
}
\examples{
# Load package
library(magrittr)

# Describe a dataset
iris \%>\% describe()

# Describe a variable
iris \%>\% describe(Species)
iris \%>\% describe(Sepal.Length)
}
