% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe_all}
\alias{describe_all}
\title{Describe all variables of a dataset}
\usage{
describe_all(data, out = "large")
}
\arguments{
\item{data}{A dataset}

\item{out}{Output format ("small"|"large")}
}
\value{
Dataset (tibble)
}
\description{
Describe all variables of a dataset
}
\examples{
describe_all(iris)
}
