% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe_cat}
\alias{describe_cat}
\title{Describe categorical variable}
\usage{
describe_cat(data, var, n, max_cat = 10, out = "text", margin = 0)
}
\arguments{
\item{data}{A dataset}

\item{var}{Variable or variable name}

\item{n}{Weights variable for count-data}

\item{max_cat}{Maximum number of categories displayed}

\item{out}{Output format ("text"|"list"|"tibble"|"df")}

\item{margin}{Left margin for text output (number of spaces)}
}
\value{
Description as text or list
}
\description{
Describe categorical variable
}
\examples{
describe_cat(iris, Species)
}
