% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe_tbl}
\alias{describe_tbl}
\title{Describe table}
\usage{
describe_tbl(data, n, target, out = "text")
}
\arguments{
\item{data}{A dataset}

\item{n}{Weights variable for count-data}

\item{target}{Target variable (binary)}

\item{out}{Output format ("text"|"list")}
}
\value{
Description as text or list
}
\description{
Describe table (e.g. number of rows and columns of dataset)
}
\examples{
describe_tbl(iris)

iris[1,1] <- NA
describe_tbl(iris)
}
