% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{describe_tbl}
\alias{describe_tbl}
\title{Describe table}
\usage{
describe_tbl(data, target, out = "text")
}
\arguments{
\item{data}{A dataset}

\item{target}{Target variable}

\item{out}{Output format ("text"|"vector")}
}
\value{
Description as text or vector
}
\description{
Describe table (e.g. number of rows and columns of dataset)
}
\examples{
describe_tbl(iris)

iris$is_virginica <- ifelse(iris$Species == "virginica", 1, 0)
describe_tbl(iris, is_virginica)
}
