% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{dwh_connect}
\alias{dwh_connect}
\title{connect to DWH}
\usage{
dwh_connect(dsn, user = NA, pwd = NA, pwd_crypt = FALSE, ...)
}
\arguments{
\item{dsn}{DSN string}

\item{user}{user name}

\item{pwd}{password of user}

\item{pwd_crypt}{is password encryption used?}

\item{...}{Further arguments to be passed to DBI::dbConnect()}
}
\value{
connection
}
\description{
connect to datawarehouse (DWH) using ODBC
}
\examples{
\dontrun{
con <- dwh_connect(dsn = "DWH1", user = "u12345")
}
}
