% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{encrypt}
\alias{encrypt}
\title{encrypt text}
\usage{
encrypt(text, codeletters = c(toupper(letters), letters, 0:9), shift = 18)
}
\arguments{
\item{text}{A text (character)}

\item{codeletters}{A string of letters that are used for encryption}

\item{shift}{Number of elements shifted}
}
\value{
Encrypted text
}
\description{
encrypt text
}
\examples{
encrypt("hello world")
}
