\name{explore-package}
\alias{explore-package}
\alias{explore}
\docType{package}
\title{
\packageTitle{explore}
}
\description{
\packageDescription{explore}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{explore}
\packageIndices{explore}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{explore}

Maintainer: \packageMaintainer{explore}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
