% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Explore-package.R
\docType{package}
\name{Explore-package}
\alias{Explore}
\alias{Explore-package}
\title{Explore: Exhaustive Procedure for LOgic-Rule Extraction (EXPLORE)}
\description{
This package is the R Implementation of the Exhaustive Procedure for LOgic-Rule Extraction (EXPLORE) in C++.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mi-erasmusmc/Explore}
}

}
\author{
\strong{Maintainer}: Aniek Markus \email{a.markus@erasmusmc.nl} (\href{https://orcid.org/0000-0001-5779-4794}{ORCID})

Authors:
\itemize{
  \item Cesar Barboza \email{c.barbozagutierrez@erasmusmc.nl} (\href{https://orcid.org/0009-0002-4453-3071}{ORCID})
}

}
\keyword{internal}
