% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Explore-package.R
\docType{package}
\name{Explore-package}
\alias{Explore}
\alias{Explore-package}
\title{Explore: Exhaustive Procedure for LOgic-Rule Extraction}
\description{
This package includes the implementation of EXPLORE in C++ and connects it to R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://cbarbozaerasmus.github.io/Explore/}
}

}
\keyword{internal}
