% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore-package.R
\docType{package}
\name{explore-package}
\alias{explore-package}
\title{explore: Simplifies Exploratory Data Analysis}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Interactive data exploration with one line of code, automated reporting or use an easy to remember set of tidy functions for low code exploratory data analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rolkra.github.io/explore/}
  \item \url{https://github.com/rolkra/explore}
  \item Report bugs at \url{https://github.com/rolkra/explore/issues}
}

}
\author{
\strong{Maintainer}: Roland Krasser \email{roland.krasser@gmail.com}

}
\keyword{internal}
