% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{explore_bar}
\alias{explore_bar}
\title{Explore categorial variable using bar charts}
\usage{
explore_bar(data, var, target, flip = TRUE, title = "", max_cat = 30,
  legend_position = "bottom", label, label_size = 2.5)
}
\arguments{
\item{data}{A dataset}

\item{var}{variable}

\item{target}{target}

\item{flip}{Should plot be flipped? (change of x and y)}

\item{title}{Title of the plot (if empty var name)}

\item{max_cat}{Maximum number of categories to be plotted}

\item{legend_position}{Position of the legend ("bottom" | "top")}

\item{label}{Show labels? (if empty, automatic)}

\item{label_size}{Size of labels}
}
\value{
Plot object (bar chart)
}
\description{
Create a barplot to explore a categorial variable.
If a target is selected, the barplot is created for all levels of the target.
}
