% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{explore_cat}
\alias{explore_cat}
\title{Explore categorial variable}
\usage{
explore_cat(data, var_cat, min_val = NA, max_val = NA, flip = TRUE,
  percent = TRUE, color = "#cccccc", auto_scale = TRUE,
  max_cat = 30)
}
\arguments{
\item{data}{A dataset}

\item{var_cat}{Name of numerical variable}

\item{min_val}{Filter values >= min_val}

\item{max_val}{Filter values <= max_val}

\item{flip}{Should plot be flipped? (change of x and y)}

\item{percent}{Plot values as percentage (instead of absolute numbers)}

\item{color}{Color of plot}

\item{auto_scale}{Not used at the moment}

\item{max_cat}{Maximum number of categories to be plotted}
}
\value{
Plot object (bar chart)
}
\description{
Create a plot to explore categorial variable
}
