% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{explore_num}
\alias{explore_num}
\title{Explore numerical variable}
\usage{
explore_num(data, var_num, min_val = NA, max_val = NA, flip = FALSE,
  color = "#cccccc", bins = 15, auto_scale = TRUE)
}
\arguments{
\item{data}{A dataset}

\item{var_num}{Name of numerical variable}

\item{min_val}{All values < min_val are converted to min_val}

\item{max_val}{All values > max_val are converted to max_val}

\item{flip}{Should plot be flipped? (change of x and y)}

\item{color}{Color of plot}

\item{bins}{Number of bins used for histogram}

\item{auto_scale}{Use 0.02 and 0.98 quantile for min_val and max_val (if min_val and max_val are not defined)}
}
\value{
Plot object (histogram)
}
\description{
Create a plot to explore numerical variable
}
