% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R, R/mapping_function_ideas.R
\name{explore_points}
\alias{explore_points}
\title{Plot spatial data frames with class "POINT"
Given a spatial data frame will produce a list of different types of visualizations.}
\usage{
explore_points(data, variable = NULL)

explore_points(data, variable = NULL)
}
\arguments{
\item{dataframe}{with geometry of "point" class}
}
\value{
A list of visualizations interactive visualizations accessible through exp. 'list$map'

A list of visualizations interactive visualizations accessible through exp. 'list$map'
}
\description{
Plot spatial data frames with class "POINT"
Given a spatial data frame will produce a list of different types of visualizations.

Plot spatial data frames with class "POINT"
Given a spatial data frame will produce a list of different types of visualizations.
}
\examples{
point_data <- st_sample(st_read(system.file("shape/nc.shp", package="sf")), 100)
point_data2 <- st_as_sf(point_data)
explore_points(point_data2)


point_data <- st_sample(st_read(system.file("shape/nc.shp", package="sf")), 100)
point_data2 <- st_as_sf(point_data)
explore_points(point_data2)


}
