% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{explore_shiny}
\alias{explore_shiny}
\title{Explore dataset interactive}
\usage{
explore_shiny(data, target)
}
\arguments{
\item{data}{A dataset}

\item{target}{Target variable (0/1 or FALSE/TRUE)}
}
\description{
Launches a shiny app to explore a dataset
}
\examples{

# Only run examples in interactive R sessions
if (interactive())  {
   explore_shiny(iris)
}
}
