% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{explore_tbl}
\alias{explore_tbl}
\title{Explore table}
\usage{
explore_tbl(data, n)
}
\arguments{
\item{data}{A dataset}

\item{n}{Weight variable for count data}
}
\description{
Explore a table. Plots variable types, variables with no variance and variables with NA
}
\examples{
explore_tbl(iris)
}
