% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{format_num}
\alias{format_num}
\title{Format number}
\usage{
format_num(number = 0, digits = 1)
}
\arguments{
\item{number}{A number (integer or real)}

\item{digits}{Number of digits}
}
\value{
Formated number as text
}
\description{
Formats a big number as k (1000) or M (100000)
}
\examples{
format_num(5500, digits = 2)
}
