% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{format_num_space}
\alias{format_num_space}
\title{Format number as character string (space as big.mark)}
\usage{
format_num_space(number = 0, digits = 1)
}
\arguments{
\item{number}{A number (integer or real)}

\item{digits}{Number of digits}
}
\value{
Formated number as text
}
\description{
Formats a big number using space as big.mark (1000 = 1 000)
}
\examples{
format_num_space(5500, digits = 2)
}
