% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{format_target}
\alias{format_target}
\title{Format target}
\usage{
format_target(target)
}
\arguments{
\item{target}{Variable as vector}
}
\value{
Formated target
}
\description{
Formats a target as a 0/1 variable
}
\examples{
iris$is_virginica <- ifelse(iris$Species == "virginica", "yes", "no")
iris$target <- format_target(iris$is_virginica)
table(iris$target)
}
