% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{format_type}
\alias{format_type}
\title{Format type description}
\usage{
format_type(type)
}
\arguments{
\item{type}{Type description ("integer", "double", "logical", character", "date")}
}
\value{
Formated type description (int|dou|log|chr|dat)
}
\description{
Format type description of varable to 3 letters (int|dou|log|chr|dat)
}
\examples{
format_type(typeof(iris$Species))
}
