% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{get_color}
\alias{get_color}
\title{get_color}
\usage{
get_color(name)
}
\arguments{
\item{name}{Name of color/color-vector}
}
\value{
Vector of color-codes
}
\description{
get_color
}
\examples{
get_color("ferrari")
}
