% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{get_nrow}
\alias{get_nrow}
\title{Get number of rows for a grid plot (deprecated, use total_fig_height() instead)}
\usage{
get_nrow(varnames, exclude = 0, ncol = 2)
}
\arguments{
\item{varnames}{List of variables to be plotted}

\item{exclude}{Number of variables that will be excluded from plot}

\item{ncol}{Number of columns (default = 2)}
}
\value{
Number of rows
}
\description{
Get number of rows for a grid plot (deprecated, use total_fig_height() instead)
}
\examples{
get_nrow(names(iris), ncol = 2)
}
