% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{get_type}
\alias{get_type}
\title{Return type of variable}
\usage{
get_type(var)
}
\arguments{
\item{var}{A vector (dataframe column)}
}
\value{
Value of typeof or "other"
}
\description{
Return value of typeof, except if variable contains <hide>, then return "other"
}
\examples{
get_type(iris$Species)
}
