% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{get_var_buckets}
\alias{get_var_buckets}
\title{Put variables into "buckets" to create a set of plots instead one large plot}
\usage{
get_var_buckets(data, bucket_size = 100, var_name_target = NA, var_name_n = NA)
}
\arguments{
\item{data}{A dataset}

\item{bucket_size}{Maximum number of variables in one bucket}

\item{var_name_target}{Name of the target variable (if defined)}

\item{var_name_n}{Name of the weight (n) variable (if defined)}
}
\value{
Buckets as a list
}
\description{
Put variables into "buckets" to create a set of plots instead one large plot
}
\examples{
get_var_buckets(iris)
get_var_buckets(iris, bucket_size = 2)
get_var_buckets(iris, bucket_size = 2, var_name_target = "Species")
}
