% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDescriptives.R
\name{getDescriptives}
\alias{getDescriptives}
\title{Explore 
 
Kurzbeschreibung
 
ausführlichere Beschreibung}
\usage{
getDescriptives(rawdata)
}
\arguments{
\item{rawdata}{data.frame input, typically a psychological dataset.}
}
\value{
list
}
\description{
Explore 
 
Kurzbeschreibung
 
ausführlichere Beschreibung
}
\examples{

}
\author{
Karl-Philipp Flösch, Joschka Cremers, Leonie Kotz
}
