% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{getSetting}
\alias{getSetting}
\title{Get the setting parameter value}
\usage{
getSetting(settings, parameter, type = "value")
}
\arguments{
\item{settings}{Settings parameter, string}

\item{parameter}{Parameter value, string}

\item{type}{Type value, string}
}
\value{
A parameter value, character.
}
\description{
Get the setting parameter value
}
