% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{guess_cat_num}
\alias{guess_cat_num}
\title{Return if variable is categorical or numerical}
\usage{
guess_cat_num(var, descr)
}
\arguments{
\item{var}{A vector (dataframe column)}

\item{descr}{A description of the variable (optional)}
}
\value{
"cat" (categorical), "num" (numerical) or "oth" (other)
}
\description{
Guess if variable is categorical or numerical based on name, type and values of variable
}
\examples{
guess_cat_num(iris$Species)
}
