% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{guess_cat_num}
\alias{guess_cat_num}
\title{Return if variable is categorial or nomerical}
\usage{
guess_cat_num(var)
}
\arguments{
\item{var}{A vector (dataframe column)}
}
\value{
"cat" (categorial), "num" (numerical) or "oth" (other)
}
\description{
Guess if variable is categorial or numerical based on name, type and values of variable
}
\examples{
guess_cat_num(iris$Species)
}
