% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{mix_color}
\alias{mix_color}
\title{Mix colors}
\usage{
mix_color(color1, color2 = NA, n = 5)
}
\arguments{
\item{color1}{Color 1}

\item{color2}{Color 2}

\item{n}{Number of different colors that should be generated}
}
\value{
Vector of color-codes
}
\description{
Mix colors
}
\examples{
mix_color("blue", n = 10)
mix_color("gold", "red", n = 4)
}
