% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunctions.R
\name{modelsCurveExplore}
\alias{modelsCurveExplore}
\title{modelsCurveExplore # TODO: update documentation?}
\usage{
modelsCurveExplore(
  train_data = NULL,
  settings_path = NULL,
  output_path,
  file_name = "train_data",
  OutputFile = NULL,
  StartRulelength = 1,
  EndRulelength = 3,
  OperatorMethod = "EXHAUSTIVE",
  CutoffMethod = "RVAC",
  ClassFeature = "'class'",
  PositiveClass = "'Iris-versicolor'",
  FeatureInclude = "",
  Maximize = "ACCURACY",
  Accuracy = 0,
  Specificity = 0,
  PrintSettings = TRUE,
  PrintPerformance = TRUE,
  Subsumption = TRUE,
  BranchBound = TRUE,
  Parallel = FALSE
)
}
\arguments{
\item{train_data}{Train data}

\item{settings_path}{A string declaring the path to the settings}

\item{output_path}{A string declaring the path to the settings}

\item{file_name}{A string declaring the the path to the file name}

\item{...}{List of arguments}
}
\value{
models for different sensitivities/specificities
}
\description{
modelsCurveExplore # TODO: update documentation?
}
