% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{plot_text}
\alias{plot_text}
\title{Plot a text}
\usage{
plot_text(text = "hello world", size = 1.2, color = "black", ggplot = FALSE)
}
\arguments{
\item{text}{Text as string}

\item{size}{Text-size}

\item{color}{Text-color}

\item{ggplot}{return a ggplot-object? (or base plot)}
}
\value{
Plot
}
\description{
Plots a text (base plot) and let you choose text-size and color
}
\examples{
plot_text("hello", size = 2, color = "red")
}
