% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunctions.R
\name{predictExplore}
\alias{predictExplore}
\title{Prediction with Explore}
\usage{
predictExplore(model, test_data)
}
\arguments{
\item{model}{Model to run predictions}

\item{test_data}{Test data}
}
\description{
`predictExplore()` executes a prediction using Explore
}
