% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{replace_na_with}
\alias{replace_na_with}
\title{Replace NA}
\usage{
replace_na_with(data, var_name, with)
}
\arguments{
\item{data}{A dataframe}

\item{var_name}{Name of variable where NAs are replaced}

\item{with}{Value instead of NA}
}
\value{
Updated dataframe
}
\description{
Replace NA values of an attribute in a dataframe
}
\examples{
data <- data.frame(nr = c(1,2,3,NA,NA))
replace_na_with(data, "nr", 0)
}
