% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Generate a report of all variables}
\usage{
report(data, target, split = TRUE, output_file, output_dir)
}
\arguments{
\item{data}{A dataset}

\item{target}{Target variable (0/1 or FALSE/TRUE)}

\item{split}{Split by target? (TRUE/FALSE)}

\item{output_file}{Filename of the html report}

\item{output_dir}{Directory where to save the html report}
}
\description{
Generate a report of all variables
If target is defined, the relation to the target is reported
}
\examples{
if (rmarkdown::pandoc_available("1.12.3"))   {
  report(iris, output_dir = tempdir())
}
}
