% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{rescale01}
\alias{rescale01}
\title{Rescales a numeric variable into values between 0 and 1}
\usage{
rescale01(x)
}
\arguments{
\item{x}{numeric vector (to be rescaled)}
}
\value{
vector with values between 0 and 1
}
\description{
Rescales a numeric variable into values between 0 and 1
}
\examples{
rescale01(0:10)
}
