% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{saveData}
\alias{saveData}
\title{saveData}
\usage{
saveData(output_path, train_data, file_name)
}
\arguments{
\item{output_path}{}

\item{train_data}{}

\item{file_name}{}
}
\description{
saveData
}
