% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunctions.R
\name{settingsExplore}
\alias{settingsExplore}
\title{Writes settings in a table for Explore}
\usage{
settingsExplore(
  settings,
  output_path,
  file_name,
  train_data = NULL,
  OutputFile = NULL,
  StartRulelength = NULL,
  EndRulelength = NULL,
  OperatorMethod = NULL,
  CutoffMethod = NULL,
  ClassFeature,
  PositiveClass,
  FeatureInclude = NULL,
  Maximize = NULL,
  Accuracy = NULL,
  Specificity = NULL,
  PrintSettings = NULL,
  PrintPerformance = NULL,
  Subsumption = NULL,
  BranchBound = NULL,
  Parallel = NULL
)
}
\arguments{
\item{settings}{}

\item{output_path}{}

\item{file_name}{}

\item{train_data}{}

\item{OutputFile}{}

\item{StartRulelength}{}

\item{EndRulelength}{}

\item{OperatorMethod}{}

\item{CutoffMethod}{}

\item{ClassFeature}{}

\item{PositiveClass}{}

\item{FeatureInclude}{}

\item{Maximize}{}

\item{Accuracy}{}

\item{Specificity}{}

\item{PrintSettings}{}

\item{PrintPerformance}{}

\item{Subsumption}{}

\item{BranchBound}{}

\item{Parallel}{}
}
\value{
Settings path
}
\description{
`settingsExplore()` defines the parameters used by Explore
}
