% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunctions.R
\name{settingsExplore}
\alias{settingsExplore}
\title{Writes settings in a table for Explore}
\usage{
settingsExplore(
  settings,
  output_path,
  file_name,
  train_data = NULL,
  OutputFile = NULL,
  StartRulelength = NULL,
  EndRulelength = NULL,
  OperatorMethod = NULL,
  CutoffMethod = NULL,
  ClassFeature,
  PositiveClass,
  FeatureInclude = NULL,
  Maximize = NULL,
  Accuracy = NULL,
  Specificity = NULL,
  PrintSettings = NULL,
  PrintPerformance = NULL,
  Subsumption = NULL,
  BranchBound = NULL,
  Parallel = NULL
)
}
\arguments{
\item{settings}{Settings parameter}

\item{output_path}{A string declaring the path to the settings}

\item{file_name}{A string declaring the the path to the file name}

\item{train_data}{Train data}

\item{OutputFile}{A string declaring the path to the output file}

\item{StartRulelength}{Integer value}

\item{EndRulelength}{Integer value}

\item{OperatorMethod}{ROCREA}

\item{CutoffMethod}{RVAC}

\item{ClassFeature}{class}

\item{PositiveClass}{Such as Iris-versicolor}

\item{FeatureInclude}{Feature restriction}

\item{Maximize}{Specificity}

\item{Accuracy}{Numeric, > 0 & < 1}

\item{Specificity}{Numeric, > 0 & < 1}

\item{PrintSettings}{True or False}

\item{PrintPerformance}{True or False}

\item{Subsumption}{True or False}

\item{BranchBound}{True or False}

\item{Parallel}{True or False}
}
\value{
Settings path
}
\description{
`settingsExplore()` defines the parameters used by Explore
}
