% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunctions.R
\name{settingsExplore}
\alias{settingsExplore}
\title{Writes settings in a table for Explore}
\usage{
settingsExplore(
  settings,
  output_path,
  file_name,
  train_data = NULL,
  OutputFile = NULL,
  StartRulelength = 1,
  EndRulelength = 3,
  OperatorMethod = "EXHAUSTIVE",
  CutoffMethod = NULL,
  ClassFeature,
  PositiveClass,
  FeatureInclude = "",
  Maximize = "ACCURACY",
  Accuracy = 0,
  Specificity = 0,
  PrintSettings = "yes",
  PrintPerformance = "yes",
  Subsumption = "yes",
  BranchBound = "yes",
  Parallel = "no"
)
}
\arguments{
\item{settings}{Settings parameter}

\item{output_path}{A string declaring the path to the settings}

\item{file_name}{A string declaring the the path to the file name}

\item{train_data}{Train data}

\item{OutputFile}{A string declaring the path to the output file}

\item{StartRulelength}{Positive integer}

\item{EndRulelength}{Positive integer}

\item{OperatorMethod}{One of list with strings, e.g. list = "EXHAUSTIVE", ...}

\item{CutoffMethod}{One of list with strings, list = "RVAC", ...}

\item{ClassFeature}{String, should be name of one of columns in data train.}

\item{PositiveClass}{1 or string (?) (should be one of elements of column 'ClassFeature' in data train)}

\item{FeatureInclude}{Empty or string (should be name of one of columns in data train)}

\item{Maximize}{One of list with strings, list = "ACCURACY", ...}

\item{Accuracy}{Float 0-1 -> default = 0 (if 0, make empty = computationally more beneficial)}

\item{Specificity}{float 0-1, default = 0}

\item{PrintSettings}{True or False}

\item{PrintPerformance}{True or False}

\item{Subsumption}{True or False}

\item{BranchBound}{True or False}

\item{Parallel}{True or False}
}
\value{
Settings path
}
\description{
`settingsExplore()` defines the parameters used by Explore
}
