% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{show_color}
\alias{show_color}
\title{Show color vector as ggplot}
\usage{
show_color(color)
}
\arguments{
\item{color}{Vector of colors}
}
\value{
ggplot
}
\description{
Show color vector as ggplot
}
\examples{
show_color("gold")
show_color(c("blue", "red", "green"))
}
