% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{target_explore_cat}
\alias{target_explore_cat}
\title{Explore categorial variable + target}
\usage{
target_explore_cat(data, var, target = "target_ind", min_val = NA,
  max_val = NA, flip = TRUE, num2char = TRUE, title = NA,
  auto_scale = TRUE, na = NA, max_cat = 30,
  legend_position = "bottom")
}
\arguments{
\item{data}{A dataset}

\item{var}{Categorial variable}

\item{target}{Target variable (0/1 or FALSE/TRUE)}

\item{min_val}{All values < min_val are converted to min_val}

\item{max_val}{All values > max_val are converted to max_val}

\item{flip}{Should plot be flipped? (change of x and y)}

\item{num2char}{If TRUE, numeric values in variable are converted into character}

\item{title}{Title of plot}

\item{auto_scale}{Not used, just for compatibility}

\item{na}{Value to replace NA}

\item{max_cat}{Maximum numbers of categories to be plotted}

\item{legend_position}{Position of legend ("right"|"bottom"|"non")}
}
\value{
Plot object
}
\description{
Create a plot to explore relation between categorial variable and a binary target
}
