% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{total_fig_height}
\alias{total_fig_height}
\title{Get fig.height for RMarkdown-junk using explore_all()}
\usage{
total_fig_height(data, n, target, nvar = NA, ncol = 2, size = 3)
}
\arguments{
\item{data}{A dataset}

\item{n}{Weights variable for count data? (TRUE / MISSING)}

\item{target}{Target variable (TRUE / MISSING)}

\item{nvar}{Number of variables to plot}

\item{ncol}{Number of columns (default = 2)}

\item{size}{fig.height of 1 plot (default = 3)}
}
\value{
Number of rows
}
\description{
Get fig.height for RMarkdown-junk using explore_all()
}
\examples{
total_fig_height(iris)
total_fig_height(iris, target = Species)
total_fig_height(nvar = 5)
}
