% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunctions.R
\name{trainExplore}
\alias{trainExplore}
\title{trainExplore This function finds the best decision rule for the given dataset based using EXPLORE.}
\usage{
trainExplore(
  output_path,
  train_data = NULL,
  settings_path = NULL,
  file_name = "train_data",
  ...
)
}
\arguments{
\item{output_path}{}

\item{train_data}{}

\item{settings_path}{}

\item{file_name}{}

\item{...}{}
}
\value{

}
\description{
trainExplore This function finds the best decision rule for the given dataset based using EXPLORE.
}
